cookit-api-master/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
django-crispy-forms==1.9.2
django-extensions==3.0.9
django-slugify-processor==0.9.0
Pillow==8.2.0
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.4.1
toml==0.10.1
django-betterforms==1.2
Faker==4.18.0
django-debug-toolbar==3.1.1
psycopg2==2.8.6
django-extra-views==0.13.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
django-filter==2.4.0
gunicorn==20.0.4
djoser==2.0.1
django-utils-six==2.0